feat: create SSR header
