+++ apps/staking/components/WalletNetworkDropdownWithLocales.tsx
@@ -1,10 +1,14 @@
+'use client';
+
 import { NETWORK } from '@/lib/constants';
 import WalletNetworkDropdown from '@session/wallet/components/WalletNetworkDropdown';
 import { useTranslations } from 'next-intl';
+import { useWallet } from '@session/wallet/hooks/wallet-hooks';
 
 export function WalletNetworkDropdownWithLocales({ className }: { className?: string }) {
+  const { isConnected } = useWallet();
   const dictionary = useTranslations('wallet.networkDropdown');
-  return (
+  return isConnected ? (
     <WalletNetworkDropdown
       className={className}
       labels={{
@@ -19,5 +23,5 @@ export function WalletNetworkDropdownWithLocales({ className }: { className?: st
       }}
       variant="outline"
     />
-  );
+  ) : null;
 }
