+++ packages/wallet/components/WalletModalButton.tsx
@@ -1,6 +1,8 @@
+'use client';
+
 import { ButtonProps } from '@session/ui/ui/button';
 import { useWeb3Modal, useWeb3ModalState } from '@web3modal/wagmi/react';
-import { WALLET_STATUS, useWallet } from '../hooks/wallet-hooks';
+import { useWallet, WALLET_STATUS } from '../hooks/wallet-hooks';
 import { useWalletButton } from '../providers/wallet-button-provider';
 import { WalletButton } from './WalletButton';
